Redo HISTFILE changes
On the Mac, the session stuff made setting HISTFILE useless because it
sets it again at the first prompt so added tests to not set HISTFILE
figuring the Apple session stuff was preferred (instead of turning it
off).  Also took some time to clean up what appears to be some archaic
settings.